Move login and profile routes into separate files
